backend/requirements.txt
numpy==1.19.0
tensorflow-gpu==2.2.0
matplotlib==3.2.2